+++ example/schema-relay/QueryRoot.js
@@ -6,30 +6,34 @@ import {
 } from 'graphql'
 
 import knex from './database'
-import joinMonster from '../../src/index'
 import User from './User'
 import Sponsor from './Sponsor'
 import { nodeField } from './Node'
 
+import joinMonster from '../../src/index'
+const options = {
+  minify: process.env.MINIFY == 1
+}
+
 export default new GraphQLObjectType({
-  description: 'global query object',
-  name: 'Query',
-  fields: () => ({
-    version: {
-      type: GraphQLString,
-      resolve: () => joinMonster.version
-    },
-    node: nodeField,
-    users: {
-      type: new GraphQLList(User),
-      resolve: (parent, args, context, ast) => {
-        return joinMonster(ast, context, sql => {
-          // place the SQL query in the response headers. ONLY for debugging. Don't do this in production
+description: 'global query object',
+name: 'Query',
+fields: () => ({
+  version: {
+    type: GraphQLString,
+    resolve: () => joinMonster.version
+  },
+  node: nodeField,
+  users: {
+    type: new GraphQLList(User),
+    resolve: (parent, args, context, ast) => {
+      return joinMonster(ast, context, sql => {
+        // place the SQL query in the response headers. ONLY for debugging. Don't do this in production
           if (context) {
             context.set('X-SQL-Preview', sql.replace(/\n/g, '%0A'))
           }
           return knex.raw(sql)
-        })
+        }, options)
       }
     },
     user: {
@@ -49,7 +53,7 @@ export default new GraphQLObjectType({
             context.set('X-SQL-Preview', sql.replace(/\n/g, '%0A'))
           }
           return knex.raw(sql)
-        })
+        }, options)
       }
     },
     sponsors: {
@@ -60,7 +64,7 @@ export default new GraphQLObjectType({
           knex.raw(sql)
           .then(data => done(null, data))
           .catch(done)
-        })
+        }, options)
       }
     }
   })